requirements.txt
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.0.4
django-import-export==1.0.0
et-xmlfile==1.0.1
gunicorn==19.7.1
jdcal==1.3
odfpy==1.3.6
openpyxl==2.5.1
pytz==2018.3
PyYAML==3.12
tablib==0.12.1
unicodecsv==0.14.1
whitenoise==3.3.1
xlrd==1.1.0
xlwt==1.3.0
psycopg2==2.7.2
